backend/requirements.txt
fastapi
uvicorn
pandas
pymongo
python-dotenv
